+++ src/OneBlinkForm.tsx
@@ -292,6 +292,25 @@ function OneBlinkForm({
     ],
   )
 
+  const obFormContainerHTMLElementRef = React.useRef<HTMLDivElement>(null)
+  React.useEffect(() => {
+    const obFormContainerHTMLElement = obFormContainerHTMLElementRef.current
+    if (obFormContainerHTMLElement) {
+      console.log(
+        'Setting toast notifications to be appended to HTML Element',
+        obFormContainerHTMLElement,
+      )
+      bulmaToast.setDoc({
+        // @ts-expect-error hack until this happens: https://github.com/rfoel/bulma-toast/issues/107
+        createElement: (...args) => document.createElement(...args),
+        body: {
+          appendChild: <T extends Node>(child: T) =>
+            obFormContainerHTMLElement.appendChild(child),
+        },
+      })
+    }
+  }, [])
+
   const handleSubmit = React.useCallback(
     (event) => {
       event.preventDefault()
@@ -301,7 +320,7 @@ function OneBlinkForm({
       if (pagesValidation) {
         bulmaToast.toast({
           message: 'Please fix validation errors',
-          // @ts-expect-error ???
+          // @ts-expect-error bulma sets this string as a class, so we are hacking in our own classes
           type: 'ob-toast is-danger cypress-invalid-submit-attempt',
           position: 'bottom-right',
           duration: 4000,
@@ -437,7 +456,7 @@ function OneBlinkForm({
   }
 
   return (
-    <div className="ob-form-container">
+    <div className="ob-form-container" ref={obFormContainerHTMLElementRef}>
       <form
         name="obForm"
         className={`ob-form cypress-ob-form ob-form__page-${